+++ CHANGELOG.rst
@@ -29,6 +29,7 @@ Change Log
   containing integer (representing the "order" on which the lines are disconnected due to protections) rather
   than just boolean.
 - [FIXED]: some bugs in the `action_space.get_all_unitary_redispatch` and `action_space.get_all_unitary_curtail`
+- [FIXED]: some bugs in the `GreedyAgent` and `TopologyGreedy`
 - [ADDED] support for the "alarm operator" / "attention budget" feature
 - [ADDED] retrieval of the `max_step` (ie the maximum number of step that can be performed for the current episode)
   in the observation